move tests to correct subdir
